refactor: moved helper functions separately.
